Changing the map now updates player args. SteamIDs are now uint64 rather than uint32.

+++ Game/LobbyInfo.cs
@@ -33,7 +33,7 @@ namespace Game
     {
         public int LobbyIndex;
         public string Name;
-        public uint SteamID;
+        public UInt64 SteamID;
         public int GamePlayer;
         public int GameTeam;
         public bool Host;
