+++ SteamWrapper/SteamWrapper.h
@@ -21,7 +21,7 @@ namespace SteamWrapper
 		static void Update();
 
 		static String^ PlayerName();
-		static UInt32 PlayerId();
+		static UInt64 PlayerId();
 	};
 
 	public ref class SteamTextInput
@@ -139,7 +139,7 @@ namespace SteamWrapper
 		static Action< bool >^ s_OnJoinLobby;
 		static Action< bool >^ s_OnCreateLobby;
 		static Action^ s_OnChatUpdate;
-		static Action< String^, uint32, String^ >^ s_OnChatMsg;
+		static Action< String^, uint64, String^ >^ s_OnChatMsg;
 		static Action^ s_OnDataUpdate;
 
 		static int s_nLobbiesFound = 0;
@@ -155,9 +155,9 @@ namespace SteamWrapper
 		
 		static int const NumLobbies();
 		static String^ GetLobbyData( int Index, String^ Key );
-		static void JoinCreatedLobby(            Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint32, String^ >^ OnChatMsg, Action^ OnDataUpdate );
-		static void JoinLobby( int Index,        Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint32, String^ >^ OnChatMsg, Action^ OnDataUpdate );
-		static void JoinLobby( CSteamID LobbyID, Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint32, String^ >^ OnChatMsg, Action^ OnDataUpdate );
+		static void JoinCreatedLobby(            Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint64, String^ >^ OnChatMsg, Action^ OnDataUpdate );
+		static void JoinLobby( int Index,        Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint64, String^ >^ OnChatMsg, Action^ OnDataUpdate );
+		static void JoinLobby( CSteamID LobbyID, Action< bool >^ OnJoinLobby, Action^ OnChatUpdate, Action< String^, uint64, String^ >^ OnChatMsg, Action^ OnDataUpdate );
 
 		static void CreateLobby(Action< bool >^ OnCreateLobby, int LobbyType );
 
@@ -172,7 +172,7 @@ namespace SteamWrapper
 
 		static int GetLobbyMemberCount();
 		static String^ GetMememberName( int Index );
-		static UInt32 GetMememberId(int Index);
+		static UInt64 GetMememberId(int Index);
 
 		static bool IsLobbyOwner();
